agro-sentinel/cloud/requirements.txt
google-cloud-pubsub==2.18.0
google-cloud-bigquery==3.11.0
google-cloud-firestore==2.11.0
google-cloud-storage==2.10.0
functions-framework==3.4.0
google-cloud-aiplatform==1.35.0
